requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-jose[cryptography]
reportlab
python-multipart
alembic
pytest
requests
